+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `pyrgg.engines.erdos_reyni` module
+- `save_log` function
 ### Changed
+- `logger` function format for `erdos_reyni_gilbert` changed
 - GitHub actions are limited to the `dev` and `master` branches
 - `README.md` modified
 - `build_exe.bat` modified